value__ is uint32_t, use %lu to print

+++ src/ats/gnunet-service-ats_performance.c
@@ -261,7 +261,7 @@ peerinfo_it (void *cls,
     return;
 
   GNUNET_log (GNUNET_ERROR_TYPE_DEBUG,
-              "Callback for peer `%s' plugin `%s' BW out %llu, BW in %llu \n",
+              "Callback for peer `%s' plugin `%s' BW out %lu, BW in %lu \n",
               GNUNET_i2s (id),
               plugin_name,
               ntohl (bandwidth_out.value__),
@@ -405,23 +405,23 @@ req_addr_peerinfo_it (void *cls,
       return;
   }
   GNUNET_log (GNUNET_ERROR_TYPE_DEBUG,
-              "Callback for  %s peer `%s' plugin `%s' BW out %u, BW in %u \n",
+              "Callback for  %s peer `%s' plugin `%s' BW out %lu, BW in %lu \n",
               (active == GNUNET_YES) ? "ACTIVE" : "INACTIVE",
               GNUNET_i2s (id),
               plugin_name,
-              (unsigned int) ntohl (bandwidth_out.value__),
-              (unsigned int) ntohl (bandwidth_in.value__));
+              ntohl (bandwidth_out.value__),
+              ntohl (bandwidth_in.value__));
 
   /* Transmit result */
   if ((GNUNET_YES == ai->all) || (GNUNET_YES == active))
   {
       GNUNET_log (GNUNET_ERROR_TYPE_DEBUG,
-                  "Sending result for  %s peer `%s' plugin `%s' BW out %u, BW in %u \n",
+                  "Sending result for  %s peer `%s' plugin `%s' BW out %lu, BW in %lu \n",
                   (active == GNUNET_YES) ? "ACTIVE" : "INACTIVE",
                   GNUNET_i2s (id),
                   plugin_name,
-                  (unsigned int) ntohl (bandwidth_out.value__),
-                  (unsigned int) ntohl (bandwidth_in.value__));
+                  ntohl (bandwidth_out.value__),
+                  ntohl (bandwidth_in.value__));
     transmit_req_addr (cls,
         id,
         plugin_name,
